+++ src/pages/Admin/CategoriesList/ThirdLevelCategory.jsx
@@ -2,27 +2,37 @@ import React from 'react';
 import { BsDot } from 'react-icons/bs';
 import { RiDeleteBin6Line } from 'react-icons/ri';
 import categoryService from '../../../services/categories';
-import { toast } from 'react-toastify';
 import UpdateCategory from './UpdateCategory';
+import Swal from 'sweetalert2';
 
-const ThirdLevelCategory = ({ setLoading, topCategoryId, secondCategoryId, thirdCat, fetchCategories }) => {
+const ThirdLevelCategory = ({ topCategoryId, secondCategoryId, thirdCat, fetchCategories }) => {
 
-  const deleteThirdLevelCategory = async () => {
-    try {
-      setLoading(true);
-      const res = await categoryService.deleteThirdLevelCategory({
-        topCategoryId,
-        secondCategoryId,
-        thirdCategoryId: thirdCat._id,
-      });
-      fetchCategories();
-      toast.error(res.message)
-    } catch (error) {
-      toast.error(error)
-    } finally {
-      setLoading(false);
-    }
+  const deleteThirdLevelCategory = () => {
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You want to delete this category!",
+      showCancelButton: true,
+      showLoaderOnConfirm: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes",
+      cancelButtonText: "No",
+      preConfirm: async () => {
+        try {
+          await categoryService.deleteThirdLevelCategory({
+            topCategoryId,
+            secondCategoryId,
+            thirdCategoryId: thirdCat._id,
+          });
+          fetchCategories();
+        } catch (error) {
+          Swal.showValidationMessage(`Request failed: ${error}`);
+        }
+      },
+      allowOutsideClick: () => !Swal.isLoading(),
+    });
   }
+
 
   return (
     <div className='subSubCategory'>
