+++ src/pages/Admin/ProductList/ProductList.jsx
@@ -3,10 +3,11 @@ import './ProductList.scss';
 import { toast } from 'react-toastify';
 import { RiDeleteBin6Line } from 'react-icons/ri';
 import productService from '../../../services/product';
-import { calculateStock } from '../../../Helper';
+import { calculateStock, formatNumbers } from '../../../Helper';
 import { useNavigate } from 'react-router-dom';
 import { MdOutlineModeEdit } from 'react-icons/md';
 import Toast from '../../../components/Toast/Toast';
+import Swal from 'sweetalert2';
 
 const ProductList = () => {
   const [products, setProducts] = useState([]);
@@ -29,15 +30,28 @@ const ProductList = () => {
     fetchAllProducts()
   }, [])
 
-  const handleDeleteProduct = async (id) => {
-    try {
-      await productService.deleteProduct(id);
-      toast.success('Product delete succussfully');
-      fetchAllProducts();
-    } catch (error) {
-      toast.error(error);
-    }
+  const handleDeleteProduct = () => {
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You want to delete this product!",
+      showCancelButton: true,
+      showLoaderOnConfirm: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes",
+      cancelButtonText: "No",
+      preConfirm: async () => {
+        try {
+          await productService.deleteProduct(id);
+          fetchAllProducts();
+        } catch (error) {
+          Swal.showValidationMessage(`Request failed: ${error}`);
+        }
+      },
+      allowOutsideClick: () => !Swal.isLoading(),
+    });
   }
+
 
   return (
     <div className='productsList'>
@@ -47,8 +61,8 @@ const ProductList = () => {
           <thead>
             <tr>
               <th>Image</th>
-              <th>Title</th>
-              <th>Category</th>
+              <th>Product Name</th>
+              <th>Categories</th>
               <th>Price</th>
               {/* <th>Quantity</th> */}
               <th>Stock</th>
@@ -68,8 +82,8 @@ const ProductList = () => {
                         <img src={'https://rukminim1.flixcart.com/image/612/612/xif0q/jean/h/y/g/34-jeans-bt008-laheja-original-imagqqbsfgmdhcvn.jpeg?q=70'} alt="img" />
                       </td>
                       <td>{item.title}</td>
-                      <td>{item.topLevelCategory}</td>
-                      <td>₹{item.discountedPrice}</td>
+                      <td>{item.topLevelCategory}/{item.secondLevelCategory}/{item.thirdLevelCategory}</td>
+                      <td>₹{formatNumbers(item.discountedPrice)}</td>
                       {/* <td>{item.quantity}</td> */}
                       <td>{calculateStock(item.size)}</td>
                       <td>
